Address review: component-level warning, IncludeWithWarning test, XML docs, extract test file
