+++ src/MicroElements.OpenApi.FluentValidation/FluentValidation/FluentValidationSchemaBuilder.cs
@@ -76,7 +76,8 @@ namespace MicroElements.OpenApi.FluentValidation
                                         logger.LogDebug("Rule '{RuleName}' applied for property '{SchemaTypeName}.{SchemaPropertyName}'", rule.Name, schemaTypeName, schemaPropertyName);
 
                                         if (schemaGenerationOptions.ConditionalRules == ConditionalRulesMode.IncludeWithWarning
-                                            && !validationRuleContext.ValidationRule.HasNoCondition())
+                                            && (!validationRuleContext.ValidationRule.HasNoCondition()
+                                                || validationRuleContext.ValidationRule.Components.Any(c => !c.HasNoCondition())))
                                         {
                                             logger.LogWarning(
                                                 "Conditional validation rule '{RuleName}' included in schema for '{SchemaTypeName}.{SchemaPropertyName}'. The .When()/.Unless() condition cannot be represented in OpenAPI schema.",
